+++ src/main/java/it/unibo/coffebreak/impl/view/renders/entityrenders/barrel/BarrelRender.java
@@ -36,7 +36,7 @@ public class BarrelRender extends AbstractEntityRender {
         super.render(g, entity);
         g.setColor(Color.BLUE);
         g.fillOval((int) entity.getPosition().x(), (int) entity.getPosition().y(), 
-                    getScaledEntityWidth(entity), getScaledEntityHeight(entity));
+                    (int) getScaledEntityWidth(entity), (int) getScaledEntityHeight(entity));
     }
 
     /**
