+++ src/main/java/it/unibo/coffebreak/impl/view/renders/entityrenders/mario/PlayerRender.java
@@ -33,7 +33,7 @@ public class PlayerRender extends AbstractEntityRender {
         super.render(g, entity);
         g.setColor(Color.GREEN);
         g.fillRect((int) entity.getPosition().x(), (int) entity.getPosition().y(), 
-                    getScaledEntityWidth(entity), getScaledEntityHeight(entity));
+                    (int) getScaledEntityWidth(entity), (int) getScaledEntityHeight(entity));
     }
 
     /**
